orchestrator/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
psycopg2-binary==2.9.9
ollama==0.1.7
python-multipart==0.0.6